server/requirements.txt
aiofiles==23.1.0
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.6.0
autopep8==2.0.2
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
class-registry==2.1.2
click==8.1.6
colorama==0.4.3
colorlog==6.7.0
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-api-logger==1.1.12
drf-yasg==1.21.5
exceptiongroup==1.1.2
fastapi==0.100.0
fastjsonschema==2.17.1
filters==1.3.2
filters-django==1.0.5
Flask==2.1.3
Flask-Cors==4.0.0
flask-sock==0.6.0
gunicorn==20.1.0
h11==0.14.0
html5tagger==1.3.0
httptools==0.6.0
humanize==4.8.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
lxml==4.9.3
Markdown==3.4.1
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
PyJWT==2.7.0
pyshorteners==1.0.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
reactpy==1.0.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
sanic==23.3.0
Sanic-Cors==2.2.0
sanic-routing==22.8.0
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
starlette==0.27.0
tomli==2.0.1
tornado==6.3.2
tracerite==1.1.0
typing_extensions==4.7.1
ujson==5.8.0
uritemplate==4.1.1
urllib3==2.0.2
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.1.2
whitenoise==6.4.0
wsproto==1.2.0
xlrd==2.0.1
zipp==3.16.0
